+++ internal/web/job/remote_routing_job.go
@@ -6,9 +6,8 @@ import (
 	"github.com/mhsanaei/3x-ui/v3/internal/web/service"
 )
 
-// RemoteRoutingJob keeps permanent Happ and Clash/Mihomo routing URLs warm.
-// Network work runs in cron (and once asynchronously at panel startup), never
-// in a subscription request handler.
+// RemoteRoutingJob keeps remote Happ and Clash/Mihomo routing URLs warm: all
+// network work runs here (cron + startup warm), never in a request handler.
 type RemoteRoutingJob struct {
 	settingService service.SettingService
 }
